pyproject.toml
[project]
name = "labor-research-platform"
version = "0.1.0"
description = "Integrated labor relations research platform: OLMS unions, F-7 employers, NLRB elections, OSHA violations, WHD wage theft, corporate ownership"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "psycopg2-binary>=2.9.11",
    "pandas>=2.3.0",
    "numpy>=2.4.0",
    "splink>=4.0.12",
    "rapidfuzz>=3.14.0",
    "cleanco>=2.3",
    "requests>=2.32.0",
    "beautifulsoup4>=4.14.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.15.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests that take >10 seconds",
    "integration: marks tests that require a running database",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]  # line length handled separately
